FIXED SOME BUGS AND RECODED SERVICES AND CONTROLLERS

+++ src/main/java/com/strong/BloodDonation/Service/MedicalHistoryService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import com.strong.BloodDonation.Model.MedicalHistory;
 import com.strong.BloodDonation.Repository.MedicalHisRepo;
+import com.strong.BloodDonation.Utils.BloodException;
 
 @Service
 public class MedicalHistoryService {
@@ -14,26 +15,51 @@ public class MedicalHistoryService {
     @Autowired
     private MedicalHisRepo medicalHisRepo;
 
-    public void saveHistory(MedicalHistory medicalHistory) {
-        medicalHisRepo.save(medicalHistory);
+    public void saveHistory(MedicalHistory medicalHistory) throws BloodException {
+        if (medicalHistory == null || medicalHistory.getDonor().getFirstName() == null
+                || medicalHistory.getMedicalCondition() == null) {
+            throw new BloodException("Invalid medical history data");
+        }
+        try {
+            medicalHisRepo.save(medicalHistory);
+        } catch (Exception e) {
+            throw new BloodException("Error saving medical history", e);
+        }
     }
 
     public MedicalHistory findById(Integer id) {
         return medicalHisRepo.findById(id).orElse(null);
     }
 
-    public List<MedicalHistory> findALl() {
+    public List<MedicalHistory> findAll() {
         return medicalHisRepo.findAll();
     }
 
-    public void deleteHistory(Integer histroyId) {
-        MedicalHistory history = findById(histroyId);
-        medicalHisRepo.delete(history);
+    public void deleteHistory(Integer historyId) throws BloodException {
+        MedicalHistory history = findById(historyId);
+        if (history != null) {
+            try {
+                medicalHisRepo.delete(history);
+            } catch (Exception e) {
+                throw new BloodException("Error deleting medical history: " + e.getLocalizedMessage());
+            }
+        } else {
+            throw new BloodException("Medical history not found with ID: " + historyId);
+        }
     }
 
-    public void updateHistory(MedicalHistory medicalHistory) {
+    public void updateHistory(MedicalHistory medicalHistory) throws BloodException {
         if (medicalHistory.getHistoryId() != null) {
-            medicalHisRepo.save(medicalHistory);
+            if (medicalHistory.getDonor() == null || medicalHistory.getMedicalCondition() == null) {
+                throw new BloodException("Invalid medical history data");
+            }
+            try {
+                medicalHisRepo.save(medicalHistory);
+            } catch (Exception e) {
+                throw new BloodException("Error updating medical history : " + e.getLocalizedMessage());
+            }
+        } else {
+            throw new BloodException("Invalid Medical History ID: " + medicalHistory.getHistoryId());
         }
     }
 }
